requirements.txt
psycopg2-binary
requests
python-dotenv
streamlit-folium
sqlalchemy
beautifulsoup4